fix: merged the upload tasks to avoid configuring AGP early
