+++ src/main/kotlin/com/bugsnag/android/gradle/BugsnagUploadSoSymTask.kt
@@ -3,6 +3,7 @@ package com.bugsnag.android.gradle
 import com.android.build.gradle.api.BaseVariantOutput
 import com.bugsnag.android.gradle.internal.AbstractSoMappingTask
 import com.bugsnag.android.gradle.internal.BugsnagHttpClientHelper
+import com.bugsnag.android.gradle.internal.NdkToolchain
 import com.bugsnag.android.gradle.internal.UploadRequestClient
 import com.bugsnag.android.gradle.internal.md5HashCode
 import com.bugsnag.android.gradle.internal.taskNameSuffix
@@ -17,6 +18,8 @@ import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputDirectory
 import org.gradle.api.tasks.Internal
+import org.gradle.api.tasks.Nested
+import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.StopExecutionException
 import org.gradle.api.tasks.TaskAction
@@ -24,7 +27,7 @@ import org.gradle.api.tasks.TaskProvider
 import java.io.File
 import java.net.HttpURLConnection.HTTP_NOT_FOUND
 
-abstract class BugsnagUploadSoSymTask : DefaultTask(), AndroidManifestInfoReceiver, BugsnagFileUploadTask {
+internal abstract class BugsnagUploadSoSymTask : DefaultTask(), AndroidManifestInfoReceiver, BugsnagFileUploadTask {
 
     @get:InputDirectory
     abstract val symbolFilesDir: DirectoryProperty
@@ -34,6 +37,13 @@ abstract class BugsnagUploadSoSymTask : DefaultTask(), AndroidManifestInfoReceiv
 
     @get:Input
     abstract val projectRoot: Property<String>
+
+    @get:Nested
+    abstract val ndkToolchain: Property<NdkToolchain>
+
+    @get:Input
+    @get:Optional
+    abstract val uploadType: Property<UploadType>
 
     @get:Internal
     internal abstract val uploadRequestClient: Property<UploadRequestClient>
@@ -47,9 +57,23 @@ abstract class BugsnagUploadSoSymTask : DefaultTask(), AndroidManifestInfoReceiv
     fun upload() {
         val rootDir = symbolFilesDir.asFile.get()
         logger.info("Bugsnag: Found shared object files for upload: $rootDir")
-        rootDir.walkTopDown()
-            .filter { it.isFile && it.extension == "gz" && it.length() >= VALID_SO_FILE_THRESHOLD }
-            .forEach { uploadSymbols(it) }
+
+        if (ndkToolchain.get().preferredMappingTool() == NdkToolchain.MappingTool.OBJDUMP) {
+            // uploadType == objdump
+            val abiDirs = rootDir.listFiles().filter { it.isDirectory }
+            abiDirs.forEach { abiDir ->
+                val arch = abiDir.name
+                abiDir.listFiles()
+                    .filter { it.extension == "gz" }
+                    .forEach { sharedObjectFile ->
+                        uploadObjdump(sharedObjectFile, arch)
+                    }
+            }
+        } else {
+            rootDir.walkTopDown()
+                .filter { it.isFile && it.extension == "gz" && it.length() >= VALID_SO_FILE_THRESHOLD }
+                .forEach { uploadSymbols(it) }
+        }
     }
 
     /**
@@ -96,6 +120,36 @@ abstract class BugsnagUploadSoSymTask : DefaultTask(), AndroidManifestInfoReceiv
         requestOutputFile.asFile.get().writeText(response)
     }
 
+    private fun uploadObjdump(mappingFile: File, arch: String) {
+        // a SO file may not contain debug info. if that's the case then the mapping file should be very small,
+        // so we try and reject it here as otherwise the event-worker will reject it with a 400 status code.
+        if (!mappingFile.exists() || mappingFile.length() < VALID_SO_FILE_THRESHOLD) {
+            logger.warn("Bugsnag: Skipping upload of empty/invalid mapping file: $mappingFile")
+            return
+        }
+
+        val sharedObjectName = mappingFile.nameWithoutExtension
+        val requestEndpoint = uploadType.get().endpoint(endpoint.get())
+        val soUploadKey = uploadType.get().uploadKey
+
+        val request = BugsnagMultiPartUploadRequest.from(this, requestEndpoint)
+        val manifestInfo = parseManifestInfo()
+        val mappingFileHash = mappingFile.md5HashCode()
+        val response = uploadRequestClient.get().makeRequestIfNeeded(manifestInfo, mappingFileHash) {
+            logger.lifecycle(
+                "Bugsnag: Uploading SO mapping file for $sharedObjectName ($arch) from $mappingFile"
+            )
+            request.uploadMultipartEntity(retryCount.get()) { builder ->
+                builder.addAndroidManifestInfo(manifestInfo)
+                builder.addFormDataPart(soUploadKey, mappingFile.name, mappingFile.asRequestBody())
+                builder.addFormDataPart("arch", arch)
+                builder.addFormDataPart("sharedObjectName", sharedObjectName)
+                builder.addFormDataPart("projectRoot", projectRoot.get())
+            }
+        }
+        requestOutputFile.asFile.get().writeText(response)
+    }
+
     companion object {
         private const val ENDPOINT_SUFFIX = "/ndk-symbol"
 
@@ -112,6 +166,8 @@ abstract class BugsnagUploadSoSymTask : DefaultTask(), AndroidManifestInfoReceiv
         fun register(
             project: Project,
             variant: BaseVariantOutput,
+            ndkToolchain: NdkToolchain,
+            uploadType: UploadType,
             generateTaskProvider: TaskProvider<out AbstractSoMappingTask>,
             httpClientHelperProvider: Provider<out BugsnagHttpClientHelper>,
             ndkUploadClientProvider: Provider<out UploadRequestClient>,
@@ -124,6 +180,9 @@ abstract class BugsnagUploadSoSymTask : DefaultTask(), AndroidManifestInfoReceiv
 
                 task.endpoint.set(bugsnag.endpoint)
 
+                task.uploadType.set(uploadType)
+                task.ndkToolchain.set(ndkToolchain)
+
                 task.manifestInfo.set(BugsnagManifestUuidTask.manifestInfoForOutput(project, variant))
                 task.symbolFilesDir.set(generateTaskProvider.flatMap { it.outputDirectory })
                 task.requestOutputFile.set(requestOutputFileFor(project, variant))
@@ -134,6 +193,15 @@ abstract class BugsnagUploadSoSymTask : DefaultTask(), AndroidManifestInfoReceiv
 
                 task.configureWith(bugsnag)
             }
+        }
+    }
+
+    enum class UploadType(private val path: String, val uploadKey: String) {
+        NDK("so-symbol", "soSymbolFile"),
+        UNITY("so-symbol-table", "soSymbolTableFile");
+
+        fun endpoint(base: String): String {
+            return "$base/$path"
         }
     }
 }
